WIP (refactor): move modules into packages, split bot_service and bot_config
