--- config.py
@@ -1,1 +1,0 @@
-# TODO вынести сюда константы из bot_state
